refactor: restructure rule modules (#24)
